+++ src/main/kotlin/site/siredvin/turtlematic/api/AutomataCoreTraits.kt
@@ -7,4 +7,6 @@ object AutomataCoreTraits {
     val DURABILITY_REFUND_CHANCE = ResourceLocation(Turtlematic.MOD_ID, "durability_refund_chance")
     val DURABILITY_REFUND = ResourceLocation(Turtlematic.MOD_ID, "durability_refund")
     val STARBOUND_REGENERATION = ResourceLocation(Turtlematic.MOD_ID, "starbound_regeneration")
+    val FUEL_CONSUMPTION_DISABLED = ResourceLocation(Turtlematic.MOD_ID, "fuel_consumption_disabled")
+    val SKILLED = ResourceLocation(Turtlematic.MOD_ID, "unskilled")
 }