+++ src/main/kotlin/site/siredvin/turtlematic/common/setup/Items.kt
@@ -33,8 +33,12 @@ object Items {
         BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableHusbandryAutomataCore }).register("starbound_husbandry_automata_core")
     val STARBOUND_END_AUTOMATA_CORE =
         BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableEndAutomataCore }).register("starbound_end_automata_core")
+    val CREATIVE_HUSBANDRY_AUTOMATA_CORE =
+        BaseAutomataCore(AutomataCoreTier.CREATIVE, { TurtlematicConfig.enableHusbandryAutomataCore }).register("creative_husbandry_automata_core")
+    val CREATIVE_END_AUTOMATA_CORE =
+        BaseAutomataCore(AutomataCoreTier.CREATIVE, { TurtlematicConfig.enableEndAutomataCore }).register("creative_end_automata_core")
     val ENORMOUS_AUTOMATA_CORE =
-        BaseAutomataCore(AutomataCoreTier.ENORMOUS_TIER, { TurtlematicConfig.enableEnormousAutomata }).register("enormous_automata_core")
+        BaseAutomataCore(AutomataCoreTier.CREATIVE, { TurtlematicConfig.enableEnormousAutomata }).register("enormous_automata_core")
 
     // Forged automata cores
 
@@ -48,6 +52,11 @@ object Items {
     val STARBOUND_SMITHING_AUTOMATA_CORE = BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableSmithingAutomataCore}).register("starbound_smithing_automata_core")
     val STARBOUND_ENCHANTING_AUTOMATA_CORE = BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableEnchantingAutomataCore}).register("starbound_enchanting_automata_core")
     val STARBOUND_MASON_AUTOMATA_CORE = BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableMasonAutomataCore}).register("starbound_mason_automata_core")
+
+    val CREATIVE_BREWING_AUTOMATA_CORE = BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableBrewingAutomataCore}).register("creative_brewing_automata_core")
+    val CREATIVE_SMITHING_AUTOMATA_CORE = BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableSmithingAutomataCore}).register("creative_smithing_automata_core")
+    val CREATIVE_ENCHANTING_AUTOMATA_CORE = BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableEnchantingAutomataCore}).register("creative_enchanting_automata_core")
+    val CREATIVE_MASON_AUTOMATA_CORE = BaseAutomataCore(AutomataCoreTier.TIER4, { TurtlematicConfig.enableMasonAutomataCore}).register("creative_mason_automata_core")
 
     // Progression items
     val SOUL_VIAL = SoulVial().register("soul_vial")