+++ src/main/kotlin/site/siredvin/turtlematic/computercraft/ComputerCraftProxy.kt
@@ -38,8 +38,10 @@ object ComputerCraftProxy {
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.NETHERITE_HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.NETHERITE_END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
-        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.STARBOUND_HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.ENORMOUS_AUTOMATA_CORE) { turtle, side, tier -> EnormousAutomataCorePeripheral(turtle, side, tier) })
 
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.BREWING_AUTOMATA_CORE) { turtle, side, tier -> BrewingAutomataCorePeripheral(turtle, side, tier) })
@@ -50,6 +52,10 @@ object ComputerCraftProxy {
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_MASON_AUTOMATA_CORE) { turtle, side, tier -> MasonAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_ENCHANTING_AUTOMATA_CORE) { turtle, side, tier -> EnchantingAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_SMITHING_AUTOMATA_CORE) { turtle, side, tier -> SmithingAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_BREWING_AUTOMATA_CORE) { turtle, side, tier -> BrewingAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_MASON_AUTOMATA_CORE) { turtle, side, tier -> MasonAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_ENCHANTING_AUTOMATA_CORE) { turtle, side, tier -> EnchantingAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_SMITHING_AUTOMATA_CORE) { turtle, side, tier -> SmithingAutomataCorePeripheral(turtle, side, tier) })
 
         TURTLE_UPGRADES.add(PeripheralTurtleUpgrade.dynamic(Items.SOUL_SCRAPPER) { turtle, side -> SoulScrapperPeripheral(TurtlePeripheralOwner(turtle, side)) })
         TURTLE_UPGRADES.add(CHATTER_TURTLE!!)